Infer call expressions
- Seperate errors module
